requirement.txt
Django==1.10.1
djangorestframework==3.4.6
requests==2.11.1
